performencefix | Moved Link-dialog operations to new DataLinkDialog
